[NUI] Move ItemView related classes from internal dir to test.
ItemView is not common control, only used in test sample.
This patch is to move it from internal to test.
