+++ test/Tizen.NUI.Samples/Tizen.NUI.Samples/Components/DefaultItemLayoutProperty.cs
@@ -17,11 +17,10 @@
 
 using System.ComponentModel;
 
-namespace Tizen.NUI
+namespace Tizen.NUI.Samples
 {
     /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
     [EditorBrowsable(EditorBrowsableState.Never)]
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("Naming", "CA1707: Identifiers should not contain underscores")]
     public enum DefaultItemLayoutProperty
     {
         /// <summary>
