+++ test/Tizen.NUI.Samples/Tizen.NUI.Samples/Components/DefaultItemLayoutType.cs
@@ -17,7 +17,7 @@
 
 using System.ComponentModel;
 
-namespace Tizen.NUI
+namespace Tizen.NUI.Samples
 {
     /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
     [EditorBrowsable(EditorBrowsableState.Never)]
