+++ test/Tizen.NUI.Samples/Tizen.NUI.Samples/Components/ItemIdCollection.cs
@@ -17,33 +17,32 @@
 
 using System.ComponentModel;
 
-namespace Tizen.NUI
+namespace Tizen.NUI.Samples
 {
-
     /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
     [EditorBrowsable(EditorBrowsableState.Never)]
-    public class ItemCollection : Disposable, global::System.Collections.IEnumerable,
-        global::System.Collections.Generic.IEnumerable<Item>
+    public class ItemIdCollection : Disposable, global::System.Collections.IEnumerable,
+        global::System.Collections.Generic.IList<uint>
     {
 
-        internal ItemCollection(global::System.IntPtr cPtr, bool cMemoryOwn) : base(cPtr, cMemoryOwn)
+        internal ItemIdCollection(global::System.IntPtr cPtr, bool cMemoryOwn) : base(cPtr, cMemoryOwn)
         {
         }
 
         /// This will be public opened.
         [EditorBrowsable(EditorBrowsableState.Never)]
-        protected override void ReleaseSwigCPtr(System.Runtime.InteropServices.HandleRef swigCPtr)
+        protected override void ReleaseSwigCPtr(global::System.Runtime.InteropServices.HandleRef swigCPtr)
         {
-            Interop.ItemContainer.DeleteItemContainer(swigCPtr);
+            Interop.ItemIdContainer.DeleteItemIdContainer(swigCPtr);
         }
 
         /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
         [EditorBrowsable(EditorBrowsableState.Never)]
-        public ItemCollection(global::System.Collections.ICollection c) : this()
+        public ItemIdCollection(global::System.Collections.ICollection c) : this()
         {
             if (c == null)
                 throw new global::System.ArgumentNullException(nameof(c));
-            foreach (Item element in c)
+            foreach (uint element in c)
             {
                 this.Add(element);
             }
@@ -71,7 +70,7 @@ namespace Tizen.NUI
 
         /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
         [EditorBrowsable(EditorBrowsableState.Never)]
-        public Item this[int index]
+        public uint this[int index]
         {
             get
             {
@@ -121,21 +120,21 @@ namespace Tizen.NUI
 
         /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
         [EditorBrowsable(EditorBrowsableState.Never)]
-        public void CopyTo(Item[] array)
+        public void CopyTo(uint[] array)
         {
             CopyTo(0, array, 0, this.Count);
         }
 
         /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
         [EditorBrowsable(EditorBrowsableState.Never)]
-        public void CopyTo(Item[] array, int arrayIndex)
+        public void CopyTo(uint[] array, int arrayIndex)
         {
             CopyTo(0, array, arrayIndex, this.Count);
         }
 
         /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
         [EditorBrowsable(EditorBrowsableState.Never)]
-        public void CopyTo(int index, Item[] array, int arrayIndex, int count)
+        public void CopyTo(int index, uint[] array, int arrayIndex, int count)
         {
             if (array == null)
                 throw new global::System.ArgumentNullException(nameof(array));
@@ -153,21 +152,21 @@ namespace Tizen.NUI
                 array.SetValue(getitemcopy(index + i), arrayIndex + i);
         }
 
-        global::System.Collections.Generic.IEnumerator<Item> global::System.Collections.Generic.IEnumerable<Item>.GetEnumerator()
+        global::System.Collections.Generic.IEnumerator<uint> global::System.Collections.Generic.IEnumerable<uint>.GetEnumerator()
         {
-            return new ItemCollectionEnumerator(this);
+            return new ItemIdCollectionEnumerator(this);
         }
 
         global::System.Collections.IEnumerator global::System.Collections.IEnumerable.GetEnumerator()
         {
-            return new ItemCollectionEnumerator(this);
+            return new ItemIdCollectionEnumerator(this);
         }
 
         /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
         [EditorBrowsable(EditorBrowsableState.Never)]
-        public ItemCollectionEnumerator GetEnumerator()
+        public ItemIdCollectionEnumerator GetEnumerator()
         {
-            return new ItemCollectionEnumerator(this);
+            return new ItemIdCollectionEnumerator(this);
         }
 
         // The type-safe enumerator.
@@ -177,17 +176,17 @@ namespace Tizen.NUI
         /// tricky to detect unmanaged code that modifies the collection under our feet.
         /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
         [EditorBrowsable(EditorBrowsableState.Never)]
-        public sealed class ItemCollectionEnumerator : global::System.Collections.IEnumerator,
-            global::System.Collections.Generic.IEnumerator<Item>
+        public sealed class ItemIdCollectionEnumerator : global::System.Collections.IEnumerator
+          , global::System.Collections.Generic.IEnumerator<uint>
         {
-            private ItemCollection collectionRef;
+            private ItemIdCollection collectionRef;
             private int currentIndex;
             private object currentObject;
             private int currentSize;
 
             /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
             [EditorBrowsable(EditorBrowsableState.Never)]
-            public ItemCollectionEnumerator(ItemCollection collection)
+            public ItemIdCollectionEnumerator(ItemIdCollection collection)
             {
                 collectionRef = collection;
                 currentIndex = -1;
@@ -198,7 +197,7 @@ namespace Tizen.NUI
             // Type-safe iterator Current
             /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
             [EditorBrowsable(EditorBrowsableState.Never)]
-            public Item Current
+            public uint Current
             {
                 get
                 {
@@ -208,7 +207,7 @@ namespace Tizen.NUI
                         throw new global::System.InvalidOperationException("Enumeration finished.");
                     if (currentObject == null)
                         throw new global::System.InvalidOperationException("Collection modified.");
-                    return (Item)currentObject;
+                    return (uint)currentObject;
                 }
             }
 
@@ -264,110 +263,110 @@ namespace Tizen.NUI
         [EditorBrowsable(EditorBrowsableState.Never)]
         public void Clear()
         {
-            Interop.ItemContainer.Clear(SwigCPtr);
+            Interop.ItemIdContainer.Clear(SwigCPtr);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
         /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
         [EditorBrowsable(EditorBrowsableState.Never)]
-        public void Add(Item x)
+        public void Add(uint x)
         {
-            Interop.ItemContainer.Add(SwigCPtr, Item.getCPtr(x));
+            Interop.ItemIdContainer.Add(SwigCPtr, x);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
         private uint size()
         {
-            uint ret = Interop.ItemContainer.size(SwigCPtr);
+            uint ret = Interop.ItemIdContainer.size(SwigCPtr);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
         private uint capacity()
         {
-            uint ret = Interop.ItemContainer.capacity(SwigCPtr);
+            uint ret = Interop.ItemIdContainer.capacity(SwigCPtr);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
         private void reserve(uint n)
         {
-            Interop.ItemContainer.reserve(SwigCPtr, n);
+            Interop.ItemIdContainer.reserve(SwigCPtr, n);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
         /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
         [EditorBrowsable(EditorBrowsableState.Never)]
-        public ItemCollection() : this(Interop.ItemContainer.NewItemContainer(), true)
-        {
-            if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
-        }
-
-        /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public ItemCollection(ItemCollection other) : this(Interop.ItemContainer.NewItemContainer(ItemCollection.getCPtr(other)), true)
+        public ItemIdCollection() : this(Interop.ItemIdContainer.NewItemIdContainer(), true)
         {
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
         /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
         [EditorBrowsable(EditorBrowsableState.Never)]
-        public ItemCollection(int capacity) : this(Interop.ItemContainer.NewItemContainer(capacity), true)
+        public ItemIdCollection(ItemIdCollection other) : this(Interop.ItemIdContainer.NewItemIdContainer(ItemIdCollection.getCPtr(other)), true)
         {
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
-        private Item getitemcopy(int index)
+        /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public ItemIdCollection(int capacity) : this(Interop.ItemIdContainer.NewItemIdContainer(capacity), true)
         {
-            Item ret = new Item(Interop.ItemContainer.getitemcopy(SwigCPtr, index), true);
+            if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
+        }
+
+        private uint getitemcopy(int index)
+        {
+            uint ret = Interop.ItemIdContainer.getitemcopy(SwigCPtr, index);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        private Item getitem(int index)
+        private uint getitem(int index)
         {
-            Item ret = new Item(Interop.ItemContainer.getitem(SwigCPtr, index), false);
+            uint ret = Interop.ItemIdContainer.getitem(SwigCPtr, index);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        private void setitem(int index, Item val)
+        private void setitem(int index, uint val)
         {
-            Interop.ItemContainer.setitem(SwigCPtr, index, Item.getCPtr(val));
+            Interop.ItemIdContainer.setitem(SwigCPtr, index, val);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
         /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
         [EditorBrowsable(EditorBrowsableState.Never)]
-        public void AddRange(ItemCollection values)
+        public void AddRange(ItemIdCollection values)
         {
-            Interop.ItemContainer.AddRange(SwigCPtr, ItemCollection.getCPtr(values));
+            Interop.ItemIdContainer.AddRange(SwigCPtr, ItemIdCollection.getCPtr(values));
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
         /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
         [EditorBrowsable(EditorBrowsableState.Never)]
-        public ItemCollection GetRange(int index, int count)
+        public ItemIdCollection GetRange(int index, int count)
         {
-            global::System.IntPtr cPtr = Interop.ItemContainer.GetRange(SwigCPtr, index, count);
-            ItemCollection ret = (cPtr == global::System.IntPtr.Zero) ? null : new ItemCollection(cPtr, true);
+            global::System.IntPtr cPtr = Interop.ItemIdContainer.GetRange(SwigCPtr, index, count);
+            ItemIdCollection ret = (cPtr == global::System.IntPtr.Zero) ? null : new ItemIdCollection(cPtr, true);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
         /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
         [EditorBrowsable(EditorBrowsableState.Never)]
-        public void Insert(int index, Item x)
+        public void Insert(int index, uint x)
         {
-            Interop.ItemContainer.Insert(SwigCPtr, index, Item.getCPtr(x));
+            Interop.ItemIdContainer.Insert(SwigCPtr, index, x);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
         /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
         [EditorBrowsable(EditorBrowsableState.Never)]
-        public void InsertRange(int index, ItemCollection values)
+        public void InsertRange(int index, ItemIdCollection values)
         {
-            Interop.ItemContainer.InsertRange(SwigCPtr, index, ItemCollection.getCPtr(values));
+            Interop.ItemIdContainer.InsertRange(SwigCPtr, index, ItemIdCollection.getCPtr(values));
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
@@ -375,7 +374,7 @@ namespace Tizen.NUI
         [EditorBrowsable(EditorBrowsableState.Never)]
         public void RemoveAt(int index)
         {
-            Interop.ItemContainer.RemoveAt(SwigCPtr, index);
+            Interop.ItemIdContainer.RemoveAt(SwigCPtr, index);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
@@ -383,16 +382,16 @@ namespace Tizen.NUI
         [EditorBrowsable(EditorBrowsableState.Never)]
         public void RemoveRange(int index, int count)
         {
-            Interop.ItemContainer.RemoveRange(SwigCPtr, index, count);
+            Interop.ItemIdContainer.RemoveRange(SwigCPtr, index, count);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
         /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
         [EditorBrowsable(EditorBrowsableState.Never)]
-        public static ItemCollection Repeat(Item value, int count)
+        public static ItemIdCollection Repeat(uint value, int count)
         {
-            global::System.IntPtr cPtr = Interop.ItemContainer.Repeat(Item.getCPtr(value), count);
-            ItemCollection ret = (cPtr == global::System.IntPtr.Zero) ? null : new ItemCollection(cPtr, true);
+            global::System.IntPtr cPtr = Interop.ItemIdContainer.Repeat(value, count);
+            ItemIdCollection ret = (cPtr == global::System.IntPtr.Zero) ? null : new ItemIdCollection(cPtr, true);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
@@ -401,7 +400,7 @@ namespace Tizen.NUI
         [EditorBrowsable(EditorBrowsableState.Never)]
         public void Reverse()
         {
-            Interop.ItemContainer.Reverse(SwigCPtr);
+            Interop.ItemIdContainer.Reverse(SwigCPtr);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
@@ -409,16 +408,52 @@ namespace Tizen.NUI
         [EditorBrowsable(EditorBrowsableState.Never)]
         public void Reverse(int index, int count)
         {
-            Interop.ItemContainer.Reverse(SwigCPtr, index, count);
+            Interop.ItemIdContainer.Reverse(SwigCPtr, index, count);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
         /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
         [EditorBrowsable(EditorBrowsableState.Never)]
-        public void SetRange(int index, ItemCollection values)
+        public void SetRange(int index, ItemIdCollection values)
         {
-            Interop.ItemContainer.SetRange(SwigCPtr, index, ItemCollection.getCPtr(values));
+            Interop.ItemIdContainer.SetRange(SwigCPtr, index, ItemIdCollection.getCPtr(values));
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
+        }
+
+        /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public bool Contains(uint value)
+        {
+            bool ret = Interop.ItemIdContainer.Contains(SwigCPtr, value);
+            if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
+            return ret;
+        }
+
+        /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public int IndexOf(uint value)
+        {
+            int ret = Interop.ItemIdContainer.IndexOf(SwigCPtr, value);
+            if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
+            return ret;
+        }
+
+        /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public int LastIndexOf(uint value)
+        {
+            int ret = Interop.ItemIdContainer.LastIndexOf(SwigCPtr, value);
+            if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
+            return ret;
+        }
+
+        /// This will be public opened in next tizen after ACR done. Before ACR, need to be hidden as inhouse API.
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public bool Remove(uint value)
+        {
+            bool ret = Interop.ItemIdContainer.Remove(SwigCPtr, value);
+            if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
+            return ret;
         }
     }
 }
